+++ Model/MenuItem.cs
@@ -4,7 +4,7 @@
     {
         public int MenuItemId { get; set; }
         public string Naam { get; set; }
-        public decimal Prijs { get; set; }
+        public double Prijs { get; set; }
         public bool IsAlcoholisch { get; set; }
         public int MenuId { get; set; }
         public int Voorraad { get; set; }
